Move functions that are not HF-specific out of `hf_wrapper`
These functions relied on HF transformers in the past, but have become
independent in past refactorings, so move them to `transformer_model`
package.

Since `hf_wrapper` only contained the loader, rename it to `hf_loader`
for symmetry with the `tokenization` package layout.
