+++ curated_transformers/tests/models/test_hf_model.py
@@ -1,6 +1,5 @@
 from typing import Callable
 from dataclasses import dataclass
-from functools import partial
 from curated_transformers.models.albert import AlbertEncoder
 from curated_transformers.models.albert.config import AlbertConfig
 from curated_transformers.models.bert import BertConfig, BertEncoder
@@ -11,7 +10,8 @@ import torch
 from torch.nn import Module
 
 # fmt: off
-from curated_transformers.models.hf_wrapper import build_hf_transformer_encoder_v1, build_hf_encoder_loader_v1
+from curated_transformers.models.hf_loader import build_hf_encoder_loader_v1
+from curated_transformers.models.transformer_model import _pytorch_encoder
 from curated_transformers._compat import has_hf_transformers, transformers
 # fmt: on
 
@@ -33,7 +33,7 @@ TEST_MODELS = [
 
 def encoder_from_config(config: ModelConfig):
     encoder = config.encoder(config.config)
-    model = build_hf_transformer_encoder_v1(encoder)
+    model = _pytorch_encoder(encoder)
     model.init = build_hf_encoder_loader_v1(name=config.hf_model_name)
     return model
